rollback to mysql with bug fix on local cache expiration.

+++ s2core/src/main/scala/com/daumkakao/s2graph/core/models/ServiceColumn.scala
@@ -12,12 +12,15 @@ object ServiceColumn {
   def findById(id: Int, useCache: Boolean = true): ServiceColumn = {
     Model.find[ServiceColumn](useCache)(Seq(("id" -> id))).get
   }
+
   def find(serviceId: Int, columnName: String, useCache: Boolean = true): Option[ServiceColumn] = {
     Model.find[ServiceColumn](useCache)(Seq("serviceId" -> serviceId, "columnName" -> columnName))
   }
+
   def findsByServiceId(serviceId: Int, useCache: Boolean = true): List[ServiceColumn] = {
     Model.findsMatch[ServiceColumn](useCache)(Seq("serviceId" -> serviceId))
   }
+
   def findOrInsert(serviceId: Int,
                    columnName: String,
                    columnType: Option[String],
@@ -34,16 +37,19 @@ object ServiceColumn {
     }
   }
 }
+
 case class ServiceColumn(kvsParam: Map[KEY, VAL]) extends Model[ServiceColumn]("HServiceColumn", kvsParam) {
   override val columns = Seq("id", "serviceId", "columnName", "columnType", "schemaVersion")
   val pk = Seq(("id", kvs("id")))
   val idxServiceIdColumnName = Seq(("serviceId", kvs("serviceId")), ("columnName", kvs("columnName")))
   override val idxs = List(pk, idxServiceIdColumnName)
+
   override def foreignKeys() = {
     List(
       Model.findsMatch[ColumnMeta](useCache = false)(Seq("columnId" -> kvs("id")))
     )
   }
+
   validate(columns, Seq("schemaVersion"))
 
   val schemaVersion = kvs.get("schemaVersion").getOrElse(InnerVal.DEFAULT_VERSION).toString
@@ -55,8 +61,9 @@ case class ServiceColumn(kvsParam: Map[KEY, VAL]) extends Model[ServiceColumn]("
 
   lazy val service = Service.findById(serviceId)
   lazy val metas = ColumnMeta.findAllByColumn(id.get)
-  lazy val metasInvMap = metas.map { meta => meta.name -> meta} toMap
-  lazy val metaNamesMap = (ColumnMeta.lastModifiedAtColumn :: metas).map(x => (x.seq, x.name)) toMap
+  lazy val metasInvMap = metas.map { meta => meta.name -> meta } toMap
+  lazy val metaNamesMap = (ColumnMeta.lastModifiedAtColumn :: metas).map(x => (x.seq.toInt, x.name)) toMap
+  //  lazy val metaNamesMap = (ColumnMeta.lastModifiedAtColumn :: metas).map(x => (x.seq, x.name)) toMap
   lazy val toJson = Json.obj("serviceName" -> service.serviceName, "columnName" -> columnName, "columnType" -> columnType)
 
 }
